fix: repair portfolio cover and preview generator
