+++ public/previews/manifest.json
@@ -1,5 +1,5 @@
 {
-  "generated": "2026-08-02T13:07:39.897Z",
+  "generated": "2026-08-03T15:34:45.209Z",
   "previews": {
     "ac8bc5b8d08d": {
       "generatedAt": "2026-05-13T07:45:44.634Z",
